WIP: Refatoração cadastro e helpers
