+++ PyNFSe/cli/cadastro/__init__.py
@@ -1,0 +1,6 @@
+from .ambiente import ambiente, retornar_ambiente
+from .configuracao import configuracao, retornar_configuracao
+from .lote_rps import lote_rps
+from .prestador import prestador, retornar_prestador
+from .servico import servico
+from .tomador import tomador, retornar_tomador
