+++ PyNFSe/cli/helpers.py
@@ -1,9 +1,6 @@
-import click
 import json
 import os
 
-from PyNFSe.cli import constants
-from PyNFSe.utils.entidades import Configuracao, Prestador, Tomador, Ambiente
 from PyNFSe.nfse.pr.curitiba import NFSeCuritiba
 
 
@@ -35,36 +32,6 @@ def criar_arquivo_json(filename):
 def salvar_arquivo_json(arquivo_json, obj):
     with open(arquivo_json, mode='w') as file:
         json.dump(obj.__dict__, file, ensure_ascii=False)
-
-
-
-def retornar_tomador():
-    numero_doc = click.prompt('Numero documento')
-    with open(constants.JSON_TOMADORES, mode='r') as file:
-        clientes = json.load(file)
-    try:
-        return Tomador(**clientes[numero_doc])
-    except KeyError:
-        click.echo('{} não cadastrado'.format(numero_doc))
-
-
-def retornar_prestador():
-    with open(constants.JSON_PRESTADOR) as file:
-        return Prestador(**json.load(file))
-
-
-def retornar_configuracao():
-    with open(constants.JSON_CONFIGURACAO) as file:
-        return Configuracao(**json.load(file))
-
-
-def retornar_ambiente(producao):
-    if producao:
-        ambiente = constants.JSON_AMBIENTE_PRODUCAO
-    else:
-        ambiente = constants.JSON_AMBIENTE_HOMOLOGACAO
-    with open(ambiente) as file:
-        return Ambiente(**json.load(file))
 
 
 def enviar_lote(lote_rps, ambiente, producao):
